Create a separated constrained impersonation handler
